+++ frontend/src/admin/sections/room/view/room-view.jsx
@@ -31,11 +31,18 @@ export function RoomView() {
         if (table.selected.length === 0) return;
 
         try {
+            const jwt = localStorage.getItem('jwt');
+
+            if (!jwt) {
+                console.error('JWT token is missing');
+                return;
+            }
+
             for (const roomId of table.selected) {
                 const response = await fetch(`${import.meta.env.VITE_API_URL}/api/admin/rooms/delete/${roomId}`, {
                     method: 'DELETE',
                     headers: {
-                        "Content-Type": "application/json",
+                        'Authorization': 'Bearer ' + jwt,
                     },
                     // credentials: 'include',
                 });
@@ -57,10 +64,18 @@ export function RoomView() {
         const fetchRooms = async () => {
             setLoading(true);
             try {
+                const jwt = localStorage.getItem('jwt');
+
+                if (!jwt) {
+                    console.error('JWT token is missing');
+                    return;
+                }
+
                 const response = await fetch(`${import.meta.env.VITE_API_URL}/api/admin/rooms`, {
                     method: 'GET',
                     headers: {
                         "Content-Type": "application/json",
+                        'Authorization': 'Bearer ' + jwt,
                     },
                     // credentials: 'include',
                 });
