+++ frontend/src/admin/sections/user/user-table-row.jsx
@@ -6,10 +6,17 @@ import { Link, useNavigate } from "react-router-dom";
 
 const deleteUser = async (id) => {
     try {
+        const jwt = localStorage.getItem('jwt');
+
+        if (!jwt) {
+            console.error('JWT token is missing');
+            return;
+        }
+
         const response = await fetch(`${import.meta.env.VITE_API_URL}/api/admin/users/delete/${id}`, {
             method: 'DELETE',
             headers: {
-                "Content-Type": "application/json",
+                'Authorization': 'Bearer ' + jwt,
             },
             // credentials: 'include',
         });
